+++ CHANGES.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+###
+- Added `LuaSyntaxOptions.Luau`.
+
+### Changed
+- Changed `LuaSyntaxOptions.ToString` to return `"Luau"` for the luau preset.
 
 ## v0.2.9-beta.3
 ### Added